74X version for LHCP plots

+++ python/OurSelectionDec2012_cff.py
@@ -66,6 +66,7 @@ dimuons = cms.EDProducer('Zprime2muCompositeCandidatePicker',
                          sort_by_pt = cms.bool(True),
                          do_remove_overlap = cms.bool(True),
                          back_to_back_cos_angle_min = cms.double(-0.9998), # this corresponds to the angle (pi - 0.02) rad = 178.9 deg
-                         vertex_chi2_max = cms.double(10),
+#                         vertex_chi2_max = cms.double(10),
+                         vertex_chi2_max = cms.double(20),
                          dpt_over_pt_max = cms.double(0.3)
                          )
